\name{allNLGmm-class}
\Rdversion{1.1}
\docType{class}
\alias{allNLGmm-class}

\title{Class \code{"allNLGmm"}}
\description{
Non linear classes for GMM models, with subclasses \code{"functionGmm"}
and \code{"nonlinearGmm"}.  
}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
No methods defined with class "allNLGmm" in the signature.
}

\keyword{classes}
