\name{gmmModels-class}
\Rdversion{1.1}
\docType{class}
\alias{gmmModels-class}

\title{Class \code{"gmmModels"}}
\description{
A class for all gmm Models
}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
  \describe{
    \item{print}{\code{signature(x = "gmmModels")}: ... }
    \item{show}{\code{signature(object = "gmmModels")}: ... }
	 }
}

\examples{
showClass("gmmModels")
}
\keyword{classes}
