\name{print-methods}
\docType{methods}
\alias{print-methods}
\alias{print,ANY-method}
\alias{print,gmmModels-method}
\alias{print,gelModels-method}
\alias{print,gmmfit-method}
\alias{print,gelfit-method}
\alias{print,sgmmfit-method}
\alias{print,rlinearGmm-method}
\alias{print,rslinearGmm-method}
\alias{print,rfunctionGmm-method}
\alias{print,rformulaGmm-method}
\alias{print,rnonlinearGmm-method}
\alias{print,summaryGmm-method}
\alias{print,summarySysGmm-method}
\alias{print,specTest-method}
\alias{print,gmmWeights-method}
\alias{print,sysGmmWeights-method}
\alias{print,sysGmmModels-method}
\alias{print,hypothesisTest-method}
\title{ Methods for Function \code{print} in Package \pkg{base}}
\description{
Print method for all \code{"gmmModels"}, \code{"gmmfit"},
\code{"summaryGmm"} \code{"hypothesisTest"} and \code{"specTest"}  objects. 
}
\section{Methods}{
\describe{

\item{\code{signature(x = "ANY")}}{
}
\item{\code{signature(x = "gmmModels")}}{
}
\item{\code{signature(x = "gelModels")}}{
}

\item{\code{signature(x = "sysGmmModels")}}{
}
\item{\code{signature(x = "sysGmmWeights")}}{
}
\item{\code{signature(x = "gmmfit")}}{
}
\item{\code{signature(x = "gelfit")}}{
}
\item{\code{signature(x = "sgmmfit")}}{
}
\item{\code{signature(x = "summaryGmm")}}{
}
\item{\code{signature(x = "summarySysGmm")}}{
}
\item{\code{signature(x = "specTrest")}}{
}

\item{\code{signature(x = "rlinearGmm")}}{
}

\item{\code{signature(x = "rformulaGmm")}}{
}

\item{\code{signature(x = "rslinearGmm")}}{
}

\item{\code{signature(x = "rnonlinearGmm")}}{
}

\item{\code{signature(x = "rfunctionGmm")}}{
}

\item{\code{signature(x = "hypothesisTest")}}{
}

\item{\code{signature(x = "gmmWeights")}}{
}

}}

\keyword{methods}
