\name{regGmm-class}
\docType{class}
\alias{regGmm-class}

\title{Class \code{"regGmm"}}
\description{
Regression classes for GMM models, with subclasses \code{"linearGmm"}
and \code{"nonlinearGmm"}.
}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Extends}{
Class \code{"\linkS4class{gmmModels}"}, directly.
}
\section{Methods}{
  \describe{
    \item{evalMoment}{\code{signature(object = "regGmm")}: ... }
	 }
}

\keyword{classes}
