\name{Restricted-GEL}
\docType{class}
\alias{rgelModels-class}

\title{Classes for Restricted GEL models}
\description{
  \code{"rgelModels"} is a union class for all restricted GEL
  models. 
}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
  \describe{
    \item{modelFit}{\code{signature(object = "rgelModels")}: ... }
	 }
}
\seealso{
See \code{\linkS4class{rlinearGel}}, \code{\linkS4class{rnonlinearGel}}
and \code{\linkS4class{rfunctionGel}} for the restricted model classes.
}
\examples{
showClass("rgelModels")
}
\keyword{classes}
