\name{rsysGmmModels-class}
\docType{class}
\alias{rsysGmmModels-class}

\title{Class \code{"rsysGmmModels"}}
\description{
A union class that includes all restricted system of equations.
}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
No methods defined with class "rsysGmmModels" in the signature.
}

\examples{
showClass("rsysGmmModels")
}
\keyword{classes}
