\name{sysGmmModels-class}
\docType{class}
\alias{sysGmmModels-class}

\title{Class \code{"sysGmmModels"}}
\description{
  A class for all system GMM Models
}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
  \describe{
    \item{print}{\code{signature(x = "sysGmmModels")}: ... }
    \item{show}{\code{signature(object = "sysGmmModels")}: ... }
	 }
}

\examples{
showClass("sysGmmModels")
}
\keyword{classes}
