\name{union-class}
\docType{class}
\alias{numericORNULL-class}
\alias{numericORlogical-class}
\alias{callORNULL-class}
\alias{functionORNULL-class}
\alias{numericORcharacter-class}
\alias{matrixORnumeric-class}
\alias{matrixORcharacter-class}
\alias{numericORmatrixORNULL-class}
\alias{expressionORNULL-class}

\title{Some Union Classes}
\description{
A set of union classes for some methods.
}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
No methods defined for those classes in the signature.
}

\keyword{classes}
