\name{Restricted-GMM}
\docType{class}
\alias{rgmmModels-class}

\title{Class \code{"rgmmModels"}}
\description{
  A union class for all restricted GEL models.
}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
  \describe{
    \item{modelFit}{\code{signature(object = "rgmmModels")}: ... }
	 }
}
\seealso{
See \code{\linkS4class{rlinearGmm}} \code{\linkS4class{rnonlinearGmm}}
for the restricted model classes.
}
\examples{
showClass("rgmmModels")
}
\keyword{classes}
